Move babel_selector to babel module
